Refaktoroi: Erottele suppeiden tietojen käsittely selkeämmin;
Tämän ansiosta kaikki suppeisiin tietoihin wrappääminen tehdään yhdessä paikassa, jolloin
muu koodi pysyy selkeämpänä, kun siellä käsitellään aina pelkkiä laajoja tyyppejä. Muun koodin
refaktorointi edelleen helpottuu samalla.
