+++ src/main/scala/fi/oph/koski/valpas/rouhinta/ValpasKuntarouhintaService.scala
@@ -21,7 +21,7 @@ class ValpasKuntarouhintaService(application: KoskiApplication)
     with Logging
     with GlobalExecutionContext
 {
-  private val oppijaService = application.valpasOppijaService
+  private val oppijaLaajatTiedotService = application.valpasOppijaLaajatTiedotService
   private val rouhintaOvKeskeytyksetService = application.valpasRouhintaOppivelvollisuudenKeskeytysService
 
   def haeKunnanPerusteellaIlmanOikeustarkastusta
@@ -123,7 +123,7 @@ class ValpasKuntarouhintaService(application: KoskiApplication)
     val oppivelvollisetKoskessa = getOppivelvollisetKotikunnalla(kunta)
 
     rouhintaTimed("haeKunnanPerusteellaKoskesta", oppivelvollisetKoskessa.size) {
-      oppijaService
+      oppijaLaajatTiedotService
         // Kunnan käyttäjällä on aina oikeudet kaikkiin oppijoihin, joilla on oppivelvollisuus voimassa, joten
         // käyttöoikeustarkistusta ei tarvitse tehdä
         .getOppijalistaIlmanOikeustarkastusta(oppivelvollisetKoskessa)
@@ -145,7 +145,7 @@ class ValpasKuntarouhintaService(application: KoskiApplication)
 
   private def getOppivelvollisetKotikunnalla(kunta: String): Seq[ValpasHenkilö.Oid] = {
     timed("getOppivelvollisetKotikunnalla") {
-      oppijaService.getOppivelvollisetKotikunnallaIlmanOikeustarkastusta(kunta).map(_.masterOid)
+      oppijaLaajatTiedotService.getOppivelvollisetKotikunnallaIlmanOikeustarkastusta(kunta).map(_.masterOid)
     }
   }
 }
